+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implemented `Hash` and `Eq` for `EntityId`.
+
+### Changed
+- Rounds can now be initiated by multiple actors.
 
 ## [0.6.0] - 2020-03-11
 ### Added